refactor: make unified <Header> component
