--- project/src/components/header-room/header-room.tsx
@@ -1,39 +1,0 @@
-import { UserLogin } from 'src/types/types';
-import { Link } from 'react-router-dom';
-
-type HeaderRoomProps = {
-  userLogin: UserLogin;
-}
-
-export default HeaderRoom;
-
-function HeaderRoom({ userLogin }: HeaderRoomProps): JSX.Element {
-  return (
-    <header className="header">
-      <div className="container">
-        <div className="header__wrapper">
-          <div className="header__left">
-            <Link className="header__logo-link" to="/">
-              <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
-            </Link>
-          </div>
-          <nav className="header__nav">
-            <ul className="header__nav-list">
-              <li className="header__nav-item user">
-                <div className="header__nav-profile">
-                  <div className="header__avatar-wrapper user__avatar-wrapper"></div>
-                  <span className="header__user-name user__name">{userLogin}</span>
-                </div>
-              </li>
-              <li className="header__nav-item">
-                <a className="header__nav-link" href="#">
-                  <span className="header__signout">Sign out</span>
-                </a>
-              </li>
-            </ul>
-          </nav>
-        </div>
-      </div>
-    </header>
-  );
-}
